requirements.txt
google-adk
requests
json5
pydantic
asyncio
a2a-sdk
google-genai
anthropic
crewai
stagehand
stagehand-py
python-a2a